Basic game complete
2p game functions are complete including win message and general gameplay. No AI as of yet,
but I intend to come back to complete these parts and redesign the error message system.
Overall, the design could be better and the presentation of player data could be improved, but
my goal here was more to focus on using the module pattern to use minimal global variables, and
given that I have ultimately only used 4 lines of global code, I would say that the goal has been met.